refactor(BaseCommentParser): rollout new base class
