+++ src/test/suite/cairo-parser/getScopeParsingResult/constructor.test.ts
@@ -3,11 +3,11 @@ import * as path from "path";
 import * as fs from "fs";
 import CairoParser from "../../../../lib/CairoParser";
 import FunctionSignatureRegexParser from "../../../../lib/parser/function-signature/regex";
-import FunctionCommentDescParser from "../../../../lib/parser/function-comment-new/desc";
-import FunctionCommentImplicitArgsParser from "../../../../lib/parser/function-comment-new/implicit-args";
-import FunctionCommentExplicitArgsParser from "../../../../lib/parser/function-comment-new/explicit-args";
-import FunctionCommentReturnsParser from "../../../../lib/parser/function-comment-new/returns";
-import FunctionCommentRaisesParser from "../../../../lib/parser/function-comment-new/raises";
+import FunctionCommentDescParser from "../../../../lib/parser/function-comment/desc";
+import FunctionCommentImplicitArgsParser from "../../../../lib/parser/function-comment/implicit-args";
+import FunctionCommentExplicitArgsParser from "../../../../lib/parser/function-comment/explicit-args";
+import FunctionCommentReturnsParser from "../../../../lib/parser/function-comment/returns";
+import FunctionCommentRaisesParser from "../../../../lib/parser/function-comment/raises";
 
 suite("getScopeParsingResult: constructor", () => {
   test("should get `1` for the length of function scope", () => {
