+++ src/test/suite/function-comment/constructor/explicit-args.test.ts
@@ -1,7 +1,7 @@
 import * as assert from "assert";
 import * as path from "path";
 import * as fs from "fs";
-import FunctionCommentExplicitArgsParser from "../../../../lib/parser/function-comment-new/explicit-args";
+import FunctionCommentExplicitArgsParser from "../../../../lib/parser/function-comment/explicit-args";
 import CairoParser from "../../../../lib/CairoParser";
 
 suite("function-comment: constructor: explicit-args", () => {
